(trunk) Merging 3.2 Residential and other enhancements and bug fixes into Trunk.
git-svn-id: https://svn.code.sf.net/p/gridlab-d/code/trunk@5481 54b63ff4-7b51-0410-8bda-9fa96797e505

+++ pthreads/create.c
@@ -142,7 +142,7 @@ pthread_create (pthread_t * tid,
 
   if (a != NULL)
     {
-      stackSize = a->stacksize;
+      stackSize = (long)a->stacksize;
       tp->detachState = a->detachstate;
       priority = a->param.sched_priority;
